# openapi.json
{"openapi":"3.1.0","info":{"title":"BookManagementSyatem","summary":"Help to reommand books based on review and rating","description":"This app is used to manage books","version":"0.0.1"},"paths":{"/users/me":{"get":{"summary":"Read Current User","operationId":"read_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBasic":[]}]}},"/books":{"get":{"summary":"Get All Books","operationId":"get_all_books_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBasic":[]}]},"post":{"summary":"Create Book","operationId":"create_book_books_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestBook"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBasic":[]}]}},"/books/{book_id}":{"get":{"summary":"Get Book","operationId":"get_book_books__book_id__get","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Book","operationId":"update_book_books__book_id__put","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestBook"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Book","operationId":"delete_book_books__book_id__delete","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book_id}/reviews":{"get":{"summary":"Get Book Reviews","operationId":"get_book_reviews_books__book_id__reviews_get","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Add Book Review","operationId":"add_book_review_books__book_id__reviews_post","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestReview"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book_id}/summary":{"get":{"summary":"Get Book Summery","operationId":"get_book_summery_books__book_id__summary_get","security":[{"HTTPBasic":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendations":{"get":{"summary":"Get Recommendations","operationId":"get_recommendations_recommendations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBasic":[]}]}},"/generate-summary":{"post":{"summary":"Generate Summary","operationId":"generate_summary_generate_summary_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBasic":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RequestBook":{"properties":{"title":{"type":"string","title":"Title"},"author":{"type":"string","title":"Author"},"genre":{"type":"string","title":"Genre"},"year_published":{"type":"integer","title":"Year Published"}},"type":"object","required":["title","author","genre","year_published"],"title":"RequestBook"},"RequestReview":{"properties":{"review_text":{"type":"string","title":"Review Text"},"rating":{"type":"integer","title":"Rating"}},"type":"object","required":["review_text","rating"],"title":"RequestReview"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBasic":{"type":"http","scheme":"basic"}}}}